vector-store: extract partition_id mod from table
